pyproject.toml
[project]
name = "robotkirby"
version = "0.1.0"
description = "Robot Kirby is a Discord bot that can imitate specific users, channels, or servers, based on a Markov chain model."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alt-profanity-check>=1.7.1",
    "hikari==2.0.0",
    "hikari-tanjun==2.17.7",
    "markovify>=0.9.4",
    "matplotlib>=3.10.5",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pymongo>=4.14.0",
    "seaborn>=0.13.2",
    "vadersentiment>=3.3.2",
    "wordcloud>=1.9.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
    "pyright>=1.1.390",
]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I", ]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pyright]
include = ["robotkirby"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
typeCheckingMode = "basic"
useLibraryCodeForTypes = true
reportMissingTypeStubs = "none"
reportMissingImports = "warning"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"
reportUnknownLambdaType = "none"
reportArgumentType = "warning"
reportReturnType = "warning"
reportAttributeAccessIssue = "warning"
reportOptionalMemberAccess = "warning"
reportGeneralTypeIssues = "warning"
